requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
langchain>=0.1.0
langchain-google-genai==0.0.5
langchain-pinecone==0.0.3
langchain-community==0.3.26
pinecone-client>=3
google-generativeai==0.3.2
unstructured==0.11.6
pypdf==3.17.1
python-docx==1.1.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
tiktoken==0.5.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
typing-inspection==0.4.1
langchainhub==0.1.21
google-cloud-documentai==3.5.0